Add more null guards to uses of prefs.parent

+++ code/modules/client/preferences.dm
@@ -327,7 +327,9 @@ GLOBAL_LIST_EMPTY(preferences_datums)
 /datum/preferences/proc/ShowChoices(mob/user)
 	if(!SSatoms.initialized)
 		to_chat(user, span_warning("Please wait for the game to do a little more setup first...!"))
-		return TRUE
+		return
+	if(!user?.client) // Without a client in control, you can't do anything.
+		return
 	if(slot_randomized)
 		load_character(default_slot) // Reloads the character slot. Prevents random features from overwriting the slot if saved.
 		slot_randomized = FALSE
@@ -2462,7 +2464,7 @@ GLOBAL_LIST_EMPTY(preferences_datums)
 							qdel(selecting_species)
 							continue
 						if (selecting_species.whitelisted)
-							if (!SSwhitelists.is_whitelisted(parent.ckey, key))
+							if (parent && !SSwhitelists.is_whitelisted(parent.ckey, key))
 								qdel(selecting_species)
 								continue
 						choose_species += key
@@ -2661,12 +2663,13 @@ GLOBAL_LIST_EMPTY(preferences_datums)
 					var/desiredfps = input(user, "Choose your desired fps.\n-1 means recommended value (currently:[RECOMMENDED_FPS])\n0 means world fps (currently:[world.fps])", "Character Preference", clientfps)  as null|num
 					if (!isnull(desiredfps))
 						clientfps = sanitize_integer(desiredfps, -1, 1000, clientfps)
-						parent.fps = (clientfps < 0) ? RECOMMENDED_FPS : clientfps
+						if(parent)
+							parent.fps = (clientfps < 0) ? RECOMMENDED_FPS : clientfps
 				if("ui")
 					var/pickedui = input(user, "Choose your UI style.", "Character Preference", UI_style)  as null|anything in sortList(GLOB.available_ui_styles)
 					if(pickedui)
 						UI_style = pickedui
-						if (parent?.mob?.hud_used)
+						if (parent?.mob.hud_used)
 							parent.mob.hud_used.update_ui_style(ui_style2icon(UI_style))
 				if("pda_style")
 					var/pickedPDAStyle = input(user, "Choose your PDA style.", "Character Preference", pda_style)  as null|anything in GLOB.pda_styles
@@ -2949,12 +2952,12 @@ GLOBAL_LIST_EMPTY(preferences_datums)
 
 				if("parallaxup")
 					parallax = WRAP(parallax + 1, PARALLAX_INSANE, PARALLAX_DISABLE + 1)
-					if (parent?.mob?.hud_used)
+					if (parent?.mob.hud_used)
 						parent.mob.hud_used.update_parallax_pref(parent.mob)
 
 				if("parallaxdown")
 					parallax = WRAP(parallax - 1, PARALLAX_INSANE, PARALLAX_DISABLE + 1)
-					if (parent?.mob?.hud_used)
+					if (parent?.mob.hud_used)
 						parent.mob.hud_used.update_parallax_pref(parent.mob)
 
 				if("ambientocclusion")
@@ -3055,7 +3058,7 @@ GLOBAL_LIST_EMPTY(preferences_datums)
 		slot_randomized = TRUE
 		real_name = pref_species.random_name(gender)
 
-	if(randomise[RANDOM_HARDCORE] && parent.mob.mind && !character_setup)
+	if(randomise[RANDOM_HARDCORE] && parent?.mob.mind && !character_setup)
 		if(can_be_random_hardcore())
 			hardcore_random_setup(character, antagonist, is_latejoiner)
 
@@ -3281,9 +3284,9 @@ GLOBAL_LIST_EMPTY(preferences_datums)
 		parent << browse(null, "window=preferences_browser")
 
 /datum/preferences/proc/can_be_random_hardcore()
-	if(parent.mob.mind.assigned_role in GLOB.command_positions) //No command staff
+	if(parent && (parent.mob.mind?.assigned_role in GLOB.command_positions)) //No command staff
 		return FALSE
-	for(var/A in parent.mob.mind.antag_datums)
+	for(var/A in parent?.mob.mind?.antag_datums)
 		var/datum/antagonist/antag
 		if(antag.get_team()) //No team antags
 			return FALSE
